ShowWise-SecurityBackend/requirements.txt
"""Requirements for Security Backend"""
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==3.0.1
